requirements.txt
flask==3.0.3
dnspython==2.7.0